refactor: Move components to submodules
Move pages and state implementations from the main comprl_web.py to
submodules.
